#13663: Rename folder moreh_dot_op_backward to moreh_dot_backward
