+++ ttnn/cpp/ttnn/operations/moreh/moreh_dot_backward/moreh_dot_backward_pybind.cpp
@@ -9,7 +9,7 @@
 #include "moreh_dot_backward.hpp"
 #include "pybind11/cast.h"
 #include "pybind11/decorators.hpp"
-#include "ttnn/operations/moreh/moreh_dot_op_backward/device/moreh_dot_backward_device_operation.hpp"
+#include "ttnn/operations/moreh/moreh_dot_backward/device/moreh_dot_backward_device_operation.hpp"
 
 namespace ttnn::operations::moreh::moreh_dot_backward {
 void bind_moreh_dot_backward_operation(py::module& module) {
